requirements.txt
flask
streamlit
pdfminer.six
python-docx
pytesseract
requests
pytest